chore: [vgf-runtime] split legacy public headers
Move the existing VGF and Session declarations out of runtime.hpp into dedicated split public headers while keeping runtime.hpp as a compatibility include.

Update old call sites, install rules, and README usage to reference the split legacy headers before the new runtime API foundation changes their contents.
